+++ test/__tests__/elements/Input-test.js
@@ -1,6 +1,4 @@
 
-jest.dontMock('../../../src/index.js');
-
 import ReactDOM  from 'react-dom';
 import React     from 'react';
 import TestUtils from 'react-addons-test-utils';
@@ -86,6 +84,17 @@ describe('Input', () => {
     let input = TestUtils.findRenderedDOMComponentWithTag(instance, 'input');
 
     expect(ReactDOM.findDOMNode(input).getAttribute('type')).toMatch('text');
+  });
+
+  it('should have value for button type', () => {
+
+    let instance = TestUtils.renderIntoDocument(
+      <Input type="button" value="Submit..."></Input>
+    );
+
+    let input = TestUtils.findRenderedDOMComponentWithTag(instance, 'input');
+
+    expect(ReactDOM.findDOMNode(input).getAttribute('value')).toMatch('Submit...');
   });
 
   it('should display Input name', () => {